לוגין עובד פיקס

+++ ‏‏ParkPull/BL/CONVERTERS/ParkingConverter.cs
@@ -34,13 +34,13 @@ namespace BL
                 parkingCode = Parking.parkingCode,
                 userCode = Parking.userCode,
                 Name = Parking.Name,
-                widthPoint = Parking.widthPoint,
-                lengthPoint = Parking.lengthPoint,
-                parkHeight = Parking.parkHeight,
-                parkLength = Parking.parkLength,
-                parkWeight = Parking.parkWeight,
-                parkWidth = Parking.parkWidth,
-                price = Parking.price
+                widthPoint = (double)Parking.widthPoint,
+                lengthPoint = (double)Parking.lengthPoint,
+                parkHeight = (double)Parking.parkHeight,
+                parkLength = (double)Parking.parkLength,
+                parkWeight = (double)Parking.parkWeight,
+                parkWidth = (double)Parking.parkWidth,
+                price = (double)Parking.price
             };
         }
     }
